+++ configs/babel.js
@@ -46,7 +46,8 @@ module.exports = function (api) {
 					forceAllTransforms: es5Standalone,
 					useBuiltIns: 'entry',
 					corejs: 3,
-					modules: false
+					modules: false,
+					loose: true
 				}
 			],
 			[
@@ -76,19 +77,11 @@ module.exports = function (api) {
 			// Stage 2
 			[require('@babel/plugin-proposal-decorators').default, false],
 			// '@babel/plugin-proposal-function-sent',
-			require('@babel/plugin-proposal-export-namespace-from').default,
-			require('@babel/plugin-proposal-numeric-separator').default,
 			// '@babel/plugin-proposal-throw-expressions',
 
 			// Stage 3
-			require('@babel/plugin-syntax-dynamic-import').default,
 			// '@babel/plugin-syntax-import-meta',
-			[require('@babel/plugin-proposal-class-properties').default, {loose: true}],
 			// '@babel/plugin-proposal-json-strings'
-
-			// Soon to be included within pre-env; include here until then
-			require('@babel/plugin-proposal-optional-chaining').default,
-			require('@babel/plugin-proposal-nullish-coalescing-operator').default,
 
 			require('babel-plugin-dev-expression'),
 			env === 'test' && !es5Standalone && require('babel-plugin-dynamic-import-node').default,