+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## unreleased
+
+* Added `private-methods` and `private-property-in-object` Babel plugins to remove warning.
+
 ## 4.0.1 (June 30, 2021)
 
 * Changed to the storybook released module version to `6.2.9` to avoid build error temporarily.